requirements.txt

# Networking part

pg8000
flask
netius
unidecode

# Scientific part

pyparsing <= 2.4.7
matplotlib
numpy
scipy
ase
pycodcif
spglib
yascheduler
xylib-py # NB apt-get install libboost-all-dev
nexusformat
